refactor: Created abstract class with shared methods between section 7 and 8 style: modified some style related to titles feat: Added missing field for category in section 8
