+++ src/app/@shared/form/category/category.component.html
@@ -1,6 +1,8 @@
 <div [formGroup]="categoryGroup" class="mb-4">
   <ng-container *ngIf="!isOther">
-    <label translate>{{ categoryGroup.get('name')?.value }}</label>
+    <div class="mb-2">
+      <label translate>{{ categoryGroup.get('name')?.value }}</label>
+    </div>
   </ng-container>
 
   <!-- description -->
