+++ src/app/@shared/form/impact-evaluation/impact-evaluation.component.html
@@ -60,7 +60,9 @@
       <!-- 7.1.1.4 (only if other selected in 7.1.1.3) -->
       <ng-container *ngIf="categories.value.includes(other)">
         <div formArrayName="optionOtherCtrl">
-          <mat-label><span translate>impactEvaluation.categoryOther</span></mat-label>
+          <div class="mb-2">
+            <span translate>impactEvaluation.categoryOther</span>
+          </div>
           <ng-container
             *ngFor="let otherCtrl of arrayControls(impactEvaluation.categories, 'optionOtherCtrl'); let i = index"
             [formGroupName]="i"
@@ -170,7 +172,7 @@
           #category
           formControlName="optionCtrl"
           multiple
-          (valueChange)="onCategoryChange($event, impactEvaluation.results)"
+          (valueChange)="onScaleChange($event, impactEvaluation.results)"
         >
           <mat-option *ngFor="let option of categoriesScale" [value]="option">
             <label>{{ option.name | translate }}</label>
