+++ src/app/@shared/form/impact-evaluation/impact-evaluation.component.ts
@@ -1,18 +1,7 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { FormArray, UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { TranslateService } from '@ngx-translate/core';
-import {
-  CATEGORIES_SCALE,
-  IMPACT_DIMENSION,
-  IMPACT_EVAL_CATEGORIES,
-  IMPACT_EVALUATION,
-  IMPACT_SCALE,
-  IMPACT_SCALE_TERM,
-  IMPACT_TYPE,
-  MOCK_CATEGORIES,
-  MOCK_CATEGORY_GROUP,
-  OTHER,
-} from '../constants';
+import { IMPACT_DIMENSION, IMPACT_EVALUATION, IMPACT_TYPE, MOCK_CATEGORIES, MOCK_CATEGORY_GROUP } from '../constants';
 import { Category, SustainableDevelopmentImpactPayload } from '../interface';
 import { MitigationActionsService } from '@app/mitigation-actions/mitigation-actions.service';
 import { AdaptationActionService } from '@app/adaptation-actions/adaptation-actions-service';
@@ -21,6 +10,7 @@ import { States } from '@app/@shared/next-state';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { AdaptationAction } from '@app/adaptation-actions/interfaces/adaptationAction';
 import { MitigationAction } from '@app/mitigation-actions/mitigation-action';
+import { ImpactFormBaseComponent } from '../impact-form-base.component';
 
 @Component({
   selector: 'app-impact-evaluation',
@@ -28,7 +18,7 @@ import { MitigationAction } from '@app/mitigation-actions/mitigation-action';
   styleUrl: './impact-evaluation.component.scss',
   standalone: false,
 })
-export class ImpactEvaluationComponent {
+export class ImpactEvaluationComponent extends ImpactFormBaseComponent {
   @Output() onComplete = new EventEmitter<boolean>();
   @Output() state = new EventEmitter<States>();
   @Input() stepper: any;
@@ -38,13 +28,8 @@ export class ImpactEvaluationComponent {
   form: UntypedFormGroup;
   loading = false;
 
-  other = OTHER;
-  impactEvalCategories = IMPACT_EVAL_CATEGORIES;
-  impactScale = IMPACT_SCALE;
-  impactScaleTerm = IMPACT_SCALE_TERM;
   impactType = IMPACT_TYPE;
   impactDimension = IMPACT_DIMENSION;
-  categoriesScale = CATEGORIES_SCALE;
   impactEvaluation = IMPACT_EVALUATION;
 
   categories = MOCK_CATEGORIES; // TODO: delete
@@ -54,7 +39,9 @@ export class ImpactEvaluationComponent {
     private formBuilder: UntypedFormBuilder,
     private translateService: TranslateService,
     private snackBar: MatSnackBar,
-  ) {}
+  ) {
+    super(formBuilder);
+  }
 
   ngOnInit() {
     if (!this.adaptation) {
@@ -68,8 +55,8 @@ export class ImpactEvaluationComponent {
     }
 
     this.createForm();
-    this.watchCategorySelection(this.impactEvaluation.categories);
-    this.watchCategorySelection(this.impactEvaluation.results);
+    this.watchOptionSelection(this.impactEvaluation.categories);
+    this.watchOptionSelection(this.impactEvaluation.results);
   }
 
   get categoryGroupsToView(): any[] {
@@ -85,25 +72,21 @@ export class ImpactEvaluationComponent {
     return this.categories.filter((cat) => selectedGroupCodes.includes(cat.category_group.code));
   }
 
-  get formArray(): FormArray {
-    return this.form.get('formArray') as FormArray;
-  }
-
   private createForm() {
     this.form = this.formBuilder.group({
       formArray: this.formBuilder.array([
         this.formBuilder.group({
           dimensionCtrl: ['', Validators.required],
           categoryGroupCtrl: ['', Validators.required],
-          optionCtrl: ['', Validators.required],
-          optionOtherCtrl: this.formBuilder.array([]),
+          optionCtrl: ['', Validators.required], // category
+          optionOtherCtrl: this.formBuilder.array([]), // category other
           categoriesCtrl: this.formBuilder.array([]),
           impactTypeCtrl: ['', Validators.required],
           pertinentCtrl: ['', Validators.required],
           relevantCtrl: ['', Validators.required],
         }),
         this.formBuilder.group({
-          optionCtrl: ['', Validators.required],
+          optionCtrl: ['', Validators.required], // category
           impactScaleCtrl: ['', Validators.required],
           impactScaleTermCtrl: ['', Validators.required],
           categoriesCtrl: this.formBuilder.array([]),
@@ -113,12 +96,6 @@ export class ImpactEvaluationComponent {
   }
 
   private updateForm() {}
-
-  private getSection(section: number) {
-    const formArray = this.form.get('formArray') as FormArray;
-    const sectionGroup = formArray.at(section) as UntypedFormGroup;
-    return sectionGroup;
-  }
 
   buildCategoryPayload() {
     let categories: Category[] = [];
@@ -191,91 +168,5 @@ export class ImpactEvaluationComponent {
           });
         },
       });
-  }
-
-  arrayControls(section: number, control: string) {
-    const sectionGroup = this.formArray.at(section) as UntypedFormGroup;
-    const categories = sectionGroup.get(control) as FormArray;
-    return categories.controls;
-  }
-
-  addAnotherOption(section: number) {
-    const sectionGroup = this.getSection(section);
-    const othersArray = sectionGroup.get('optionOtherCtrl') as FormArray;
-    this.addOtherOption(othersArray);
-  }
-
-  addOtherOption(othersArray: FormArray) {
-    othersArray.push(
-      this.formBuilder.group({
-        name: ['', [Validators.required, Validators.minLength(1), Validators.maxLength(100)]],
-        description: ['', [Validators.required, Validators.minLength(50), Validators.maxLength(600)]],
-        indicator: ['', [Validators.required]],
-        indicatorOther: ['', [Validators.maxLength(200)]],
-        baseValue: ['', Validators.maxLength(70)],
-        expectedValue: ['', Validators.maxLength(70)],
-        accumulatedValue: ['', Validators.maxLength(70)],
-      }),
-    );
-  }
-
-  removeOtherOption(index: number, section: number) {
-    const sectionGroup = this.getSection(section);
-    const othersArray = sectionGroup.get('optionOtherCtrl') as FormArray;
-    othersArray.removeAt(index);
-  }
-
-  onOtherOptionChange(event: any, section: number) {
-    const sectionGroup = this.getSection(section);
-    const othersArray = sectionGroup.get('optionOtherCtrl') as FormArray;
-    if (event.includes(this.other) && !othersArray.length) {
-      this.addOtherOption(othersArray);
-    } else if (!event.includes(this.other)) {
-      while (othersArray.length > 0) {
-        othersArray.removeAt(0);
-      }
-    }
-  }
-
-  onCategoryChange(event: any, section: number) {
-    const value = event.value;
-    const sectionGroup = this.getSection(section);
-
-    if (value === this.impactEvalCategories.SCALE) {
-      sectionGroup?.get('impactScaleCtrl')?.setValidators([Validators.required]);
-      sectionGroup?.get('impactScaleTermCtrl')?.clearValidators();
-    } else if (value === this.impactEvalCategories.SCALE_TERM) {
-      sectionGroup?.get('impactScaleCtrl')?.clearValidators();
-      sectionGroup?.get('impactScaleTermCtrl')?.setValidators([Validators.required]);
-    }
-    sectionGroup?.get('impactScaleCtrl')?.updateValueAndValidity();
-    sectionGroup?.get('impactScaleTermCtrl')?.updateValueAndValidity();
-  }
-
-  private watchCategorySelection(section: number) {
-    const sectionGroup = this.getSection(section);
-    const optionCtrl = sectionGroup.get('optionCtrl');
-    const categoriesArray = sectionGroup.get('categoriesCtrl') as FormArray;
-    optionCtrl?.valueChanges.subscribe((values: any[]) => {
-      while (categoriesArray.length > 0) {
-        categoriesArray.removeAt(0);
-      }
-
-      values
-        ?.filter((value) => value !== this.other)
-        .forEach((value) => {
-          categoriesArray.push(
-            this.formBuilder.group({
-              name: [value.name],
-              description: ['', [Validators.required, Validators.minLength(50), Validators.maxLength(600)]],
-              indicator: ['', [Validators.required]],
-              indicatorOther: ['', [Validators.maxLength(200)]],
-              baseValue: ['', Validators.maxLength(70)],
-              expectedValue: ['', Validators.maxLength(70)],
-              accumulatedValue: ['', Validators.maxLength(70)],
-            }),
-          );
-        });
-    });
   }
 }
