+++ src/app/@shared/form/transformational-change/transformational-change.component.html
@@ -12,8 +12,9 @@
       </ng-template>
 
       <!-- 8.1.1 -->
-      <label translate>impactEvaluation.vision.title</label>
-      socialTransformation
+      <div class="mb-2">
+        <label translate>impactEvaluation.vision.title</label>
+      </div>
       <mat-form-field class="field-ppcn flex-1" appearance="fill">
         <mat-label> <span translate>impactEvaluation.vision.short</span></mat-label>
         <textarea
@@ -127,7 +128,9 @@
       <!-- 8.1.4 (only if other selected in 8.1.4) -->
       <ng-container *ngIf="barriersSelect.value.includes(other)">
         <div formArrayName="optionOtherCtrl">
-          <mat-label><span translate>impactEvaluation.categoryOther</span></mat-label>
+          <div class="mb-2">
+            <span translate>impactEvaluation.categoryOther</span>
+          </div>
           <ng-container
             *ngFor="
               let otherCtrl of arrayControls(transformationalChange.identification, 'optionOtherCtrl');
@@ -185,7 +188,7 @@
       </div>
     </mat-step>
 
-    <!-- Characteristics 8.2 -->
+    <!-- Characteristics 8.2.1 -->
     <mat-step formGroupName="1" [stepControl]="formArray?.get([1])">
       <ng-template matStepLabel>
         <span translate>impactEvaluation.characteristics</span>:
@@ -193,15 +196,37 @@
       </ng-template>
 
       <!-- 8.2.1.1 -->
-      <mat-form-field class="field-ppcn" appearance="fill">
+      <mat-form-field class="field-ppcn mt-4" appearance="fill">
         <mat-label><span translate>impactEvaluation.category</span></mat-label>
+        <mat-select formControlName="categoryCtrl" multiple>
+          <mat-option [value]="transCategories.technology">
+            <label translate>impactEvaluation.select.technology</label>
+          </mat-option>
+
+          <mat-option [value]="transCategories.changeAgents">
+            <label translate>impactEvaluation.select.changeAgents</label>
+          </mat-option>
+
+          <mat-option [value]="transCategories.incentives">
+            <label translate>impactEvaluation.select.incentives</label>
+          </mat-option>
+
+          <mat-option [value]="transCategories.rules">
+            <label translate>impactEvaluation.select.rules</label>
+          </mat-option>
+        </mat-select>
+      </mat-form-field>
+
+      <!-- 8.2.1.2 -->
+      <mat-form-field class="field-ppcn" appearance="fill">
+        <mat-label><span translate>impactEvaluation.characteristic</span></mat-label>
         <mat-select
           #categories
           formControlName="optionCtrl"
           multiple
           (valueChange)="onOtherOptionChange($event, transformationalChange.processes)"
         >
-          <mat-option *ngFor="let option of processesCategories" [value]="option">
+          <mat-option *ngFor="let option of characteristicsToView" [value]="option">
             <label>{{ option.name | translate }}</label>
           </mat-option>
           <mat-option [value]="other"><span translate>general.other</span></mat-option>
@@ -209,10 +234,12 @@
         <mat-error translate> errorLabel.fieldRequired </mat-error>
       </mat-form-field>
 
-      <!-- 8.2.1.2 (only if other selected in 8.2.1.1) -->
+      <!-- 8.2.1.3 (only if other selected in 8.2.1.1) -->
       <ng-container *ngIf="categories.value.includes(other)">
         <div formArrayName="optionOtherCtrl">
-          <mat-label><span translate>impactEvaluation.categoryOther</span></mat-label>
+          <div class="mb-2">
+            <span translate>impactEvaluation.categoryOther</span>
+          </div>
           <ng-container
             *ngFor="let otherCtrl of arrayControls(transformationalChange.processes, 'optionOtherCtrl'); let i = index"
             [formGroupName]="i"
@@ -243,8 +270,6 @@
         </div>
       </ng-container>
 
-      <!-- 8.2.1.3 -->
-
       <!-- 8.2.1.4 -->
       <div formArrayName="categoriesCtrl">
         <ng-container
@@ -262,6 +287,7 @@
       </div>
     </mat-step>
 
+    <!-- Characteristics 8.2.2 -->
     <mat-step formGroupName="2" [stepControl]="formArray?.get([2])">
       <ng-template matStepLabel>
         <span translate>impactEvaluation.characteristics</span>:
@@ -271,7 +297,12 @@
       <!-- 8.2.2.1 -->
       <mat-form-field class="field-ppcn mt-4" appearance="fill">
         <mat-label><span translate>impactEvaluation.category</span></mat-label>
-        <mat-select #category formControlName="optionCtrl" multiple (valueChange)="onScaleChange($event)">
+        <mat-select
+          #category
+          formControlName="optionCtrl"
+          multiple
+          (valueChange)="onScaleChange($event, transformationalChange.processes)"
+        >
           <mat-option *ngFor="let option of categoriesScale" [value]="option">
             <label>{{ option.name | translate }}</label>
           </mat-option>
