+++ src/translations/es-CR.json
@@ -846,7 +846,11 @@
       "financialInvestment": "Financieras y de inversión",
       "marketEconomicIncentives": "Mercado/Incentivos económicos",
       "structuralInequalities": "Desigualdades estructurales",
-      "other": "Otras"
+      "other": "Otras",
+      "technology": "Tecnología",
+      "changeAgents": "Agentes de cambio",
+      "incentives": "Incentivos",
+      "rules": "Normas"
     },
     "includeImpactInfo": "Desea incluir la identificación de impactos en Desarrollo Sostenible y Cambio Tranformacional para este registro?",
     "includeImpactInfoTitle": "Incluir Identificación de Impactos",
@@ -871,6 +875,7 @@
     "optional": "Considerar que aunque la información de impactos en Desarrollo Sostenible y Cambio Tranformacional es voluntaria, si se decide completar se deben registrar todos los campos obligatorios",
     "transformationalChange": "Identificación de impactos en Cambio Transformacional",
     "general": "Identificación de impactos general",
+    "characteristic": "Características",
     "characteristics": "Características e impactos en Cambio Tranformacional",
     "popUp": "En esta sección se registrarán las contribuciones de su acción para generar un cambio estructural, sostenido, irreversible y de largo plazo en la sociedad costarricense",
     "processes": "Procesos",
